1. qty not taken in decimal in printing bill

+++ app/src/main/java/com/wepindia/pos/BillingCounterSalesActivity.java
@@ -2942,7 +2942,7 @@ public class BillingCounterSalesActivity extends WepPrinterBaseActivity implemen
             Double qty = Double.parseDouble(itemQty.getText().toString().trim());
             double rate = Double.parseDouble(itemRate.getText().toString().trim());
             double amount = Double.parseDouble(itemAmount.getText().toString().trim());
-            BillKotItem billKotItem = new BillKotItem(sno, name, qty.intValue(), rate, amount);
+            BillKotItem billKotItem = new BillKotItem(sno, name, qty, rate, amount);
             billKotItems.add(billKotItem);
             count++;
